Initial commit
We are going to build an image downloading tool. Container runtimes are
overrated. Just systemd will do, the only issue then is distribution of
binaries to an immutable system image (a la CoreOS), and automatic
updates. It should be as simple as periodically polling a server,
dowloading a new file if there is one, and restarting the systemd unit.
Let's do this!