Switch to libcurl rather than Hyper/Rustls
An all-Rust static executable is nice, but the available tools are too
low-level to be convenient. The key realisation came when bikeshedding
about the manifest format: we can do everything we need by downloading
to a local file (with curl), and then only operating on local files. We
don't need to do any http manually. And then using libcurl opens up many
new possibilities: not only convenient dealing with http/https
differences, also using the system-configured certificate store, and
also the opportunity to connect over ssh, or even a (network-mounted)
local file system. This change not only makes the implementation
cleaner, it also makes the tool at least twice as useful.

There do exist Rust bindings to libcurl, a "curl" and "curl-sys" crate,
but they both have many dependencies that we do not need. We can just
link against the C library by defining a few function signatures.

Also, now that we do link against something, we can no longer use musl.
This means we also have to take libc. But it should not be a problem in
practice: libc and libcurl promise to be ABI compatible, and they are
widely available at a standard location. In particular on CoreOS we do
not need to ship anything apart from a single executable. So the
executable will still be portable. We should even be able to verify that
building on e.g. Travis produces the same bit-by-bit identical binary as
compiling on my local machine. But that is a problem for a later time.