Do not attempt to parse uris in config
If Hyper can parse it, that is no guarantee that Curl can, and the other
way around, maybe something that Curl understands is not considered to
be valid by Hyper. So let's just remove Hyper from the loop. We can do
away with the majority of the dependencies as well. Even better!