Add CLI parser
The "docopt" crate is nice, but it brings in so many dependencies (among
others "void" and "unreachable"). I don't wan't to look into what the
licenses are, study how to respect those, and how to properly credit the
dependencies, for a dozen extra crates. Let's do it by hand in ~100
lines. There was also the "getopts" crate which has a better dependency
story, but as far as I can see does not support command-style CLI's,
only --flag-style.