Add simple Python integration test driver
It will run various Tako commands in various orders and then verify that
the programs return with a zero exit code, and that the desired files
exist. At the moment, just running "tako fetch" crashes already.

Python is nice for this because it has a webserver built into the
standard library, and its support for executing programs and inspecting
files is probably as good as it gets without being too verbose. I tried
Bash as well, but it looked ugly and hacky. The Python version has a
kind of zen to it.