Overhaul CLI parser
I considered using the "clap" crate, which is reasonably lightweight
(little dependencies) and the author appears to have put considerable
effort into making sure that it does not bloat the binary. An other
alternative would be "docopt" which is nice but pulls in a regex parser.

But in the end, with all of these parser, you still have to do much of
the validation manually. What combination of flags makes sense is
something that is very application-specific, and it is difficult to
capture that in a declarative configuration. I find it easier to use
control flow for validation, even for such a simple app as Tako.
Furthermore, a handwritten parser can provide more helpful error
messages.

So introduce a light abstraction, the ArgIter, that deals with
separating flags in "--long <arg>", "--long=<arg>", "-s <arg>", and
"-s<arg>" form, and that also interprets "--" correctly. Based on this
we can then build an argument parser on top that is based mostly on
pattern matching.