Add integration test for 'tako store'
This tests that the manifest exists, and the image is in the store under
the right filename. This does not yet verify that the manifest is
downloadable, or signed correctly.

I had to comment out the previous tests (which were still failing). Now
that we can actually generate properly signed manifests, I should be
able to add a proper manifest to the test data, but for some reason the
signature does not verify. I likely confused some keys somewhere.