Put the right secret key in tests, test fetch-parse
It was a bit of a hassle to get at the PKCS8-encoded key pair, but it
can be done by providing a deterministic random number generator. The
key is still the same (because the RNG is used to generate the seed),
but now we can get at the PKCS8-ecnoded key pair for this seed too. Then
encode that as base64 and put it in the tests.

I have not yet regenerated the "foo" test manifests with this secret
key, so the two fetch tests still fail, but the store-then-fetch part
passes now.