Opt for zero-padding the version after all
I think in the end it makes more sense to consider 1.0.0 to be
equivalent to 1.0 and 1, than to order shorter versions before longer
ones. In general, just avoid comparing versions of different numbers of
parts, but still, these things happen in practice, so we have to deal
with them in some way, and the principle of least surprise applies here.

This commit only documents the choice, it has not been implemented yet.