Implement zero padding in ordering comparison
This changes the behavior according to the principle of least surprise,
I think. 1.0 should not be less than 1.0.0.

At the same time, change the bounds check to be always inclusive. We can
still do matching of user-specified bounds of the form "1.2 <= v < 2.0",
because we have Part::Min and Part::Max that can be constructed in such
a way that versions without min and max parts can be bounded by such
versions exclusively.

To give an example: to constrain 1.2 <= v < 2.0, we require that

  1.2 <= v <= 2.0.Min

The only versions with 2.0 prefix that a user can construct, compare
greater than 2.0.Min, so effectively we have v < 2.0. For wildcards this
works similarly, the asterisk just turns into min and max. For example,
to match 2.1.*, we would require

  2.1.Min <= v <= 2.1.Max

For any version w that a user can construct, it holds that

  2.1.Min < w < 2.1.Max

so effectively the wildcard covers the entire range.