Update dependencies
I released a new version of Filebuffer that depends on Winapi 0.3, so
we don't end up with the double Winapi dependency in the lock file any
more. It is mostly a matter of ceremony, because Tako is not expected
to work on Windows anyway (in MSYS perhaps it would, if it has libcurl).

Also bump a few other dependencies. The base64 crate has a new feature
that we don't use or need, but Cargo takes the newest version anyway.