Add makefile to tweak final binary
I want to strip the binary, to produce a small executable to ship and
deploy. Because preferably I would drop this in just like you would
deploy a config file. (For my Miniserver on CoreOS plans.) Then it is
advantageous to have a small file. Also, Tako is IO-bound, so we can
afford to trade performance for size.

This is a makefile for now. I recently was convinced that Nix is a good
idea, I might port this to Nix later. But Make is widespread, let's use
it for now.