Make test script a bit more ergonomic
Suppress the noise that the http server spits out; I don't want to see
it, but suppressing it properly is not so easy. Instead, we can take
advantage of the fact that it is printed to stderr rather than stdout,
and just redirect stderr to /dev/null.

Also enable socket reuse, so we can run the tests multiple times.