Add test for fetching image
Also, if there is no candidate, do not treat that as an error.
Intentionally do nothing. Whatever is in place we don't touch, so if
there is an image we don't break anything. Do not exit with a nonzero
exit code, *unless* this was an --init run and there is no image yet.
Being unable to complete the --init is definitely an error. This is
not yet implemented though.