Make output of test script TAP compliant
The TAP output is not terribly useful in itself, but as a side effects,
the script now does not terminate early when an assertion fails; it
still runs the remaining tests. Also, it now prints details to stdout,
so silencing stderr is no longer such a big problem.

Using a context manager for this is quite nice: you can capture the test
name in advance, print it afterwards, and still handle any exceptions
that occur the the body. Python is nice :)