Update manifests in unit tests
This recomputes the signatures by just running Tako itself, so this is
assuming that all of the logic is still correct, and adding the file
sizes is the only change.

The also changed, because I don't recall which files I used to generate
them. (I think I used the readme, but it changed in the mean time.) This
time I used the test "images" in tests/images, so if there are any
changes in the future, we can use these files again, and the hashes will
not change.