Update manifest in integration tests
Again, this new manifest is produced just by 'tako store', so this is
assuming that that command is still correct. There is a reasonable case
for that: the signature just takes the byte range of what is written,
and the new things that got added are the file sizes that you can just
see in the manifest.