Add corrupted origin for testing
The manifest lists:

  * 1.0.0: size in manifest is one byte smaller than actual.
  * 1.1.0: size in manifest is one byte larger than actual.
  * 2.0.0: digest in manifest is not the actual digest.

Nonetheless, the manifest has a valid signature. This exists to simulate
a malicious mirror that serves the wrong file, but with the twist that
in the first two cases the digest in the manifest is correct, so we test
that Tako rejects files based on size, rather than testing just the
digest checking.