Add tests for malicious origins
There is one test currently failing: if the manifest is wrong, in the
sense that the digest is not for a file of the size specified in the
manifest, then Tako will still accept such an image. The best thing to
do would be to reject the manifest, as it lists an impossible entry, so
something is fishy. But once we get to the download, the manifest has
been accepted already, because it had a good signature. The only thing
we can do to be on the safe side is to take no action, and reject the
image. This can have the effect that the client is prevented forever
from updating without manual intervention, if the client later observes
a mismatching manifest. Not "fixing" the manifest, but releasing a new
version, would be the right way to go if only one manifest was observed.
If the reason for the fishy manifest was a bug, then rejecting the image
is indeed the safe thing to do. If it was due to a key compromise, then
the client is screwed anyway, and the attacker could do much more damage
by uploading an image with malicious behavior.