Note to replace Ring with alternatives
Ring's current versioning and stability policy is problematic. My
upgrade of Ring from 0.13-alpha to 0.13-alpha3 in the previous commit
broke the build on Rust 1.24, which means we would be stuck on an old
(and possibly insecure) version of Ring. So either, I would need to
backport fixes to Ring manually, or try to keep master building with
Rust 1.24, which is probably not realistic. Or we switch to something
else with better stability guarantees.