Use Ed25519 from sodiumoxide, not ring
The tests still pass, with the same seed to generate the key pair. Which
means the two libraries were generating the same keys and signatures,
that's good. Everything is still compatible there.

One thing that did change is the secret key format. Ring had its own
format for storing a key pair, but now with sodiumoxide I think I'll
just store the seed and use that as secret key. (It is shorter than the
derived secret key.) To distinguish the secret key from the private key,
I'll prefix it with "PRIVATE:" everywhere, to make it harder to
accidentally put it in a public place.

Because of this change, the integration tests are now broken; the format
of the secret key is now different. I'll update those next.