Change include public key formatted secret key
We need the public key to update a manifest, because we first verify the
manifest before updating it. This might change in the future because key
rotation needs to be supported, but still. It is also useful to save the
public key together with the secret key, so you can never lose part of
the pair.

Update the gen-key command as well to use sodiumoxide now, and print the
new key pair format.

Update the integration tests to use this new key too. They pass now.