Replace sodiumoxide with rust_sodium
They are compatible for my use case; rust_sodium is a fork of
sodiumoxide, that offerst static linking of libsodium. This is a
requirement for single-binary deployment on CoreOS, because CoreOS does
not ship libsodium.

Although this gets rid of one runtime dependency, it adds 95 crates as
build dependencies, causing a significant regression in compile time.
Also, I imaginie this is a huge pain for potential packagers, especially
those who want to be able to build offline. I think this is not a
necessity; rust_sodium could just use Curl or Wget rather than pure Rust
code to download the libsodium sources, and packages could refer to
already-existing libsodium rather than the build system downloading. So
while I am not happy with this Cargo.lock explosion, I see it as a
neccessity at the moment, that can hopefully be addressed in the future.