Remove outdated comments about binary size
The opt-level=z flag has been stabilized, but currently I aim to support
Rust 1.24 which does not accept it, so it cannot yet be the default.

As for using the system allocator to reduce binary size, Rust actually
moved to using the system allocator by default in Rust 1.32, so there is
no need to opt in to it. Compile with a recent rustc, get smaller
binaries for free.