Pin the Rust toolchain to 1.48.0
The version of rustc currently shipped by Ubuntu 18.04 LTS is 1.57.0,
but Debian stable (Bullseye) ships 1.48.0, so let's go with that to
maximize compatibility. Debian oldstable (Buster) ships 1.41, but at
some point enough is enough -- if people on that OS want to build Tako,
they should build an older version of it.