Replace sodiumoxide with ed25519_compact
Sodiumoxide is no longer maintained. (Not insecure though, just no
longer maintained.) Let's look for an alternative. ed25519_compact looks
nice, only 48 KB of dependencies, most of which are optional, versus
3 MB for ed25519-dalek. It even has an option to optimize for size.

After replacing, the tests still work, even with the hard-coded seed and
keys, so this does not introduce incompatibilities.