Delete libsodium dependency
The binary size of a stripped release build grows from 337928 bytes to
436256 bytes, so ~100 KB larger. But, you no longer need libsodium.so on
the host, which on my system is 362968 bytes, so net this saves 200 KB
on systems that don't have libsodium, which is the case on Flatcar. It
also silences rustsec zeallots who are afaid of sodiumoxide because it
is unmaintained.